Work around an apparent CoreText bug in the computation of descents
It seems that CoreText has a off-by-one error when computing descents,
but only when the font format is TrueType: In font which with 1000
design units per em, CoreText will return a descent of 199 when the
font specifies 200.

The problem has been reported to Apple. For now, I'm working around
the bug. Otherwise, most test cases would fail, which would hide other
(possibly more important) errors.