Change expected advance width for test case CFF2-1
The expectation had been generated with FreeType, at a time when
FreeType had a bug in its handling of `HVAR` tables. This bug has
been fixed in FreeType, so we now need to adjust the expected
advance width. Hopefully, we'll soon have multiple implementations
of CFF2 so we don't depend just on FreeType to figure out what
value should be correct.