Initial commit of Buffalo Gals IUP test (#30)
This font has a single glyph (‘H’) which displays examples of a
boundary IUP issue which appears to be treated differently in existing
implementations. The top of the top serifs and bottom of the bottom
serifs all have multiple points at the same Y location in the default.
If you examine the bottom left serif, you will see points 149 through
157 all share the same Y coordinate of 0. When moving from the default
0.0 to -1.0 in the Hooves axis, point 153 moves by (0,50). This now
leaves 6 untouched points that need to be interpolated. The handling of
these intermediate points (150-152 & 154-156) is different between
Apple's, FreeType's and Monotype's implementations. I believe that
Monotype's solution is the correct one. In our case, intermediate
untouched points are now half the distance in Y between the touched
points.