Add a prefix to SVG identifiers
In the generated HTML document, the values of `id` attributes need to
be globally unique; otherwise, some browsers will display a different
SVG image than what we’d like to render. To distinguish the expected
from observed SVG, we add an `OBSERVED` prefix to the identifiers in
the observed rendering.