Add more test cases to GVAR-9
In macOS 10.12.4, Apple’s CoreText library has a bug where one of the
test contours doesn’t get interpolated when the `TEST` axis is set to
+0.944444.

In FreeType, there are other problems; these have been fixed by
a not-yet-submitted patch that is currently under review. The patch
will be linked from the bug report mentioned by the test description.