Added five test fonts from Ken Lunde
All of these fonts have limited Unicode coverage and limited glyph sets, and require specific strings to be used. The 'name' table reflects Unicode as the origin, and OFL is specified as the open source license.

Two of the fonts are Variable Fonts that include the 'rvrn' (Required Variation Alternates) GSUB feature, along with shaping features ('init', 'medi', and 'fina'):

TestRVRN.ttf (TrueType)
TestRVRN-CFF2.ttf (CFF2)

Background: Environments that use the latest CoreText, HarfBuzz, and possibly USE handle the test string correctly, but some apps do not. These test fonts are useful for any implementation that doesn't use CoreText, HarfBuzz, or USE for applying features. The 'rvrn' feature kicks in when the 'opsz' setting is at 30 or below, and the following glyphs transform from a square to a circle: 'medi' form of digit two, isolated form of digit three, and 'fine' form of digit 3. 

The other three are static fonts that include the 'vert' and 'vrtr" (Vertical Alternates for Rotation) GSUB features, and are meant to test support for the latter feature:

TestVRTR-Regular.otf (name-keyed OpenType/CFF)
TestVRTR-TTF-Regular.otf (TrueType)
TestVRTR--CID-Regular.otf (CID-keyed OpenType/CFF)

Background: These fonts may be the very first implementations that include the 'vrtr' GSUB feature, so broader exposure would be helpful.